refactor: extract stripReplyFallback into shared utility
Move stripReplyFallback from message_bubble.dart to a dedicated
utils/reply_fallback.dart so it can be reused across layers without
cross-importing widget code. Apply it to search results so reply
fallback prefixes (> <@user:server>) are stripped in the room list.
