+++ lib/screens/chat_screen.dart
@@ -17,7 +17,8 @@ import '../widgets/chat/emoji_picker_sheet.dart';
 import '../widgets/chat/file_send_handler.dart';
 import '../widgets/chat/long_press_wrapper.dart';
 import '../widgets/chat/message_action_sheet.dart';
-import '../widgets/chat/message_bubble.dart' show MessageBubble, stripReplyFallback;
+import '../utils/reply_fallback.dart';
+import '../widgets/chat/message_bubble.dart' show MessageBubble;
 import '../widgets/chat/reaction_chips.dart';
 import '../widgets/chat/read_receipts.dart';
 import '../widgets/chat/search_results_body.dart';
