+++ lib/utils/reply_fallback.dart
@@ -1,0 +1,12 @@
+/// Strips the Matrix reply fallback prefix (lines starting with `> `)
+/// from a message body, returning only the actual reply text.
+String stripReplyFallback(String body) {
+  final lines = body.split('\n');
+  int i = 0;
+  while (i < lines.length && (lines[i].startsWith('> ') || lines[i] == '>')) {
+    i++;
+  }
+  // Skip the blank line after the fallback block.
+  if (i < lines.length && lines[i].isEmpty) i++;
+  return lines.sublist(i).join('\n');
+}
